reminder final

+++ fos-api/FOS/FOS.API/Controllers/RecurrenceEventController.cs
@@ -53,14 +53,19 @@ namespace FOS.API.Controllers
                 return ApiUtil<Model.Dto.RecurrenceEvent>.CreateFailResult(e.ToString());
             }
         }
-        [HttpGet]
-        [Route("Add")]
-        public ApiResponse Add([FromBody]Model.Dto.RecurrenceEvent recurrenceEvent)
+        [HttpPost]
+        [Route("AddNew")]
+        public ApiResponse AddNew([FromBody]Model.Dto.RecurrenceEvent recurrenceEvent)
         {
             try
             {
-
+                recurrenceEvent.StartTempDate = recurrenceEvent.StartDate.ToLocalTime();
                 _recurrenceEventService.AddRecurrenceEvent(_recurrenceEventDtoMapper.ToModel(recurrenceEvent));
+                if (recurrenceEvent.StartTempDate >= DateTime.Now
+                    && recurrenceEvent.StartTempDate < DateTime.Now.AddHours(1))
+                {
+                    _recurrenceEventService.RunThisTask(_recurrenceEventDtoMapper.ToModel(recurrenceEvent));
+                }
                 return ApiUtil.CreateSuccessfulResult();
 
             }
@@ -75,7 +80,16 @@ namespace FOS.API.Controllers
         {
             try
             {
+                Model.Domain.RecurrenceEvent before = _recurrenceEventService.GetById(recurrenceEvent.Id);
+                recurrenceEvent.StartTempDate = recurrenceEvent.StartDate.ToLocalTime();
+                recurrenceEvent.Version += 1;
                 _recurrenceEventService.UpdateRecurrenceEvent(_recurrenceEventDtoMapper.ToModel(recurrenceEvent));
+
+                if (recurrenceEvent.StartTempDate >= DateTime.Now
+                    && recurrenceEvent.StartTempDate < DateTime.Now.AddHours(1))
+                {
+                    _recurrenceEventService.RunThisTask(_recurrenceEventDtoMapper.ToModel(recurrenceEvent));
+                }
                 return ApiUtil.CreateSuccessfulResult();
 
             }
@@ -101,6 +115,7 @@ namespace FOS.API.Controllers
         {
             try
             {
+                if (id == 0) return ApiUtil.CreateSuccessfulResult();
 
                 _recurrenceEventService.DeleteById(id);
                 return ApiUtil.CreateSuccessfulResult();
